add git info and flag_name to coveralls send
Both are needed on GitHub actions. Coveralls seems to be doing
magic things for the other supported CIs, but for GitHub actions
that magic does not work. Instead we are supposed to supply
complete information.

That means the git info member on the API call and a flag_name
member for parallel builds.